Integrated Communities Treemap into index.py
